feat(authentication): add react website with navbar and login form
The basic website structure has been set up using ReactJS and Redux to
implement to control flow (similar to the flux pattern). A basic
bootstrap navbar has been added with a simple log in form that calls the
api to authenticate the user. Upon successful authentication, the login
form is replaced with a user details snippet and logout button.