chore(setup): swap web build tool from brunch to webpack
Babel seems to have many issues with many npm libraries and therefore
it is very difficult to include pre-built libraries for use in
javascript. Webpack is more tried and tested and has fewer issues.

The conversion has also enabled the use of `SaSS` for css compilation,
which has made styling the app cleaner.