feat(User): add user sign up form on homepage
A new Homepage Banner component has been added that includes a sign up
form if no user is logged in.

The sign up form reports any errors that are returned from the service
at the point of creating a new user. Upon success it automatically
authenticates the user and displays the user status in place of the
sign up form.

If the user is already logged in, then they simply see the user status
panel and not the sign up form.