fix(sign_up): hide sign up form on initial load when already signed in
As there is an authentication request which is handled via the server
side api, it meant that the sign up form was visible momentarily before
the authentication had completed and then it flickered off.

This issue was also present in the sign in form and had been resolved so
the same solution is used for hiding the sign up form.