feat(Profile): add public profile page for all users
A new `/profile/:username` route has been added that will display the
public status of the given username.

`react-router` has been included to handle the client side routing and
this has meant some refactoring has taken place to ensure the router
loads in the correct components.

The `App.js` component is no longer required as the `Root.js` component
contains the `react-router` code and loads in the underlying components
so it became redundant.

The `Navbar` has also become a container component and handles the
saving of the user's session to local storage as this component will
always be loaded in.

The `UserStatus` component also can now display previous status history
by passing an additional boolean prop to it.