feat(Status): allow user to toggle between statuses
If a user is viewing their profile page then they will also see another
panel that allows them to toggle between their available statuses by
clicking on the radio button icon next to the desired status.

There is also an issue with `isomorphic-fetch` not overriding the
browser cache for ajax GET requests which causes calls to the service
to sometimes come back with old data as the api is not actually called
and the response is returned from the browser cache. To solve the issue
all `GET` requests are appended with a time string to ensure the
requests are unique and therefore will not cache.