fix(state): ensure user state is fetched when changes happen
If any changes to the user's state (such as status update or
authentication changes) then all states for the user need refreshing
otherwise they won't update in the other containers.

For example, if a status update is attempted but authentication fails
then the authentication state needs updating in the Profile container
otherwise the Navbar container's authentication state becomes stale and
doesn't display the correct sign in form.

Likewise, if the user signs in whilst on their own profile page then the
Navbar container user status state does not get refreshed on the
Profile container so the user's status options does not display.

At some point it may be better to refactor these multiple triggers into
a top level function and inject the state into the container components.