feat(Status): allow user to delete inactive status option
The user's listed statuses on their profile have a delete option next
to each item and upon clicking delete it will delete the status from
the user's account. A user can only delete an inactive status so no
delete option is present for the active status as it would fail at the
api.