feat(Status): allow user to add new status
The `StatusManager` now has an input field that allows a user to enter
a new status and clicking the `Add (+)` button will submit the new
status to be added.

The help text displays appropriate information when the user types into
the input box.

The `RichInputField` component has been updated to handle `onChange`
events within the input box to handle when a user types text in. Also,
it's possible to not provide a prefix icon and it won't render. The
component can also receive a `maxLength` prop to limit the length of
text allowed within the input field.

The `buildURL` utility function has been updated to use
`performance.now()` rather than `Date().getTime()` as it offers a
higher resolution of time precision reducing potential `GET` requests
being cached (as was still happening rarely with `getTime`).