style(HomeBanner): apply styles to welcome message and data block
The `HomeBanner` jumbotron has had styles applied to make it look
neater and the `data-block` section that contains the `Sign Up` form or
the signed in user's status has also been styled.

A new default site font has been added which is `Roboto` as it is a
cleaner looking font than the default `Helvetica-Neue`.