style(Profile): apply styles to profile page
The `Profile` container has had the two main blocks styled. The
`UserStatus` component now contains the relevant styling rather than
being styled by the parent container, this means it only needs to be
styled once.

`react-custom-scrollbars` has been added to provide a nicer scrollbar
for the `StatusManager` for when there is a long list of statuses to
select from.