feat(Status): add status delete confirmation modal
When clicking delete on a status in the `StatusManager` it now pops up
a confirmation modal ensuring the user is aware of the delete action.

This will ensure that users don't accidentally delete a status.

A CSS fix has been applied to the bootstrap modal as a common bug has
caused it to push the navbar contents to the right when the modal shows
which causes a jarring behaviour.