fix(action_utility): allow buildURL to accept query params
There was an issue where the `buildURL` function could not build a
correct url if the passed url parameter contained a query parameter due
to the `?` being added at the end for caching reasons.

This has now been fixed so urls with query parameters can now be passed
in successfully.