feat(TopLists): add home page top n lists showing common stats
The home page now features 3 lists that show top 10 common stats of:

  * New Users
  * Recent Activity
  * Popular Stataz

A new `DataList` component has been created that is a generic presenter
of lists and takes a list item decorator to ensure the list content can
be custom.

This has also removed the need for the specific `FollowList` component
because it has become a `DataList`.