feat(Search): add search page to display search results
A new search page has been implemented that performs a user and status
search based on the given query and displays the results on a new
search container.

The `Status` component that is used to display the status text in many
other components has now been updated so the status text is a link to
perform a search on that status to show other users with that status.