feat(Deploy): add exrm to manage deployment
Exrm has been added as a dependency to create deployment binaries for
easier releases.

The `config/prod.exs` has been updated to ensure the Endpoint runs as a
server.

The `/rel` directory created by `exrm` has been added to `.gitignore` as
it is not appropriate to commit to the repo.