chore(Profile): ensure username is always dealt with in lowercase
The username can be passed through via a query parameter and therefore
may not be in the correct case and the username should always be treated
as lowercase.

Also, the display of the username should always use the data from the
api and not the query param as it may not be type in the correct case
that the user originally entered.