fix(date_format): create date object from atomic items
Previously, the `Format.js` util woould attempt to format a date by
taking in a string of a date and converting it to a Date object based on
the string. This would work in Chrome and Internet Explorer because
their Date implementations were not strict to the standard and were more
forgiving. However, this caused a fatal error in Firefox as it is closer
to the standard and not any date string will work.