fix(http_requests): ensure https requests have a valid cacertfile
When making a HTTP Request over https, the application requires a valid
`cacertfile` to ensure the ssl handshake is valid otherwise the request
will fail.

There is a new `cafile:` config option that needs to be configured to
point to a valid `cacertfile` on the local machine.

An example valid `cacertfile` can be found at:

`https://raw.githubusercontent.com/hexpm/hex/master/lib/hex/api/ca-bundle.crt`

To clean up the reuse of setting these options, the HTTPoison requests
have been refactored into a new helper library called `HTTPRequest`
which is a simple wrapper for HTTPoison.