refactor(Profile): remove component render functionality from container
The `Profile` container no longer holds the code that renders the
profile blocks to the page. A new `ProfileSection` component has now
been created that is a dumb component responsible for the page layout
and the container is responsible for the logic.

The `ProfileSection` refresh logic has also been updated so when
profile actions are performed, the whole profile page doesn't update if
it doesn't need to.

For example, if a status is changed, then the follow blocks don't
refresh as no change has been made to them.